Added versions with serial output
Added versions that output serial start/stop commands on pin 0 so that one version can be directly plugged into an Arduino operating as a receiver unit controlling a large digit display.  This way the micro:bits can be used as a remote control for the large digit display by referees/timekeepers.